ngày mới
ngày mới

+++ Client/Model/OrderDetail.cs
@@ -11,7 +11,7 @@
 
         public int UserId { get; set; }
 
-        public int? Count { get; set; }
+        public decimal? Count { get; set; }
 
     }
 }
